+++ src/gen/pres/table-styles.ts
@@ -8,7 +8,8 @@
 
 import { CRLF, TableStyle, XML_DECL } from '../../core-enums.js'
 import type { BorderProps, TableStyleInternal, TableStyleRegionProps } from '../../core-interfaces.js'
-import { createColorElement, genXmlColorSelection, lineWidthToEmu, resolveBorderWidth } from '../../gen-utils.js'
+import { createColorElement, genXmlColorSelection, resolveBorderWidth } from '../../gen-utils.js'
+import { lineWidthToEmu } from '../../units-internal.js'
 import { el, raw, voidEl } from '../oxml/el.js'
 
 /**